1 - WebScript to return a url for the alfresco WOPI service. This is in case of proxying issues.
Note:
The alfresco-global.properties file must also match the name for the service. Reason being that because communication is now being handled share server side,
there’re a number of host resolution issues I had to surmount. For some reason, post CheckfileInfo, the url the lool service uses defaults to the values for host that is
set in the alfresco-global.properties.